requirements.txt
Flask==1.1.2
Flask-SocketIO==4.3.1
google-cloud-speech
python-engineio==3.13.2
python-socketio==4.6.0
eventlet
boto3

